accessible/tests/mochitest/states/test_editablebody.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=454997
-->
<head>
  <title>nsIAccessible states tests of contenteditable body</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      testStates(document, 0, EXT_STATE_EDITABLE);
      testStates("p", 0, EXT_STATE_EDITABLE);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>

<body contentEditable="true">

  <a target="_blank"
     title="nsIAccessible states tests of contenteditable body"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=454997">Mozilla Bug 454997</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <p id="p">hello</p>
</body>
</html>
